LayoutTests/platform/chromium/http/tests/security/dataURL/xss-DENIED-from-data-url-in-foreign-domain-subframe-expected.txt

CONSOLE MESSAGE: Unsafe JavaScript attempt to access frame with URL http://127.0.0.1:8000/security/dataURL/xss-DENIED-from-data-url-in-foreign-domain-subframe.html from frame with URL data:text/html,<html><head><script>window.onload = function(){try {top.document.getElementById('accessMe').innerHTML = 'FAIL: Cross frame access from a data: URL on a different domain was allowed';alert('FAIL: No exception thrown.');} catch (e) {alert('PASS: Exception thrown successfully.');}if (window.layoutTestController)layoutTestController.notifyDone();}</script></head><body><p>Inner-inner iframe. This iframe (which is data: URL and whose parent is on a foreign domain) is the frame attempting to access the main frame.  It should not have access to it.</p></body></html>. Domains, protocols and ports must match.

ALERT: PASS: Exception thrown successfully.
The scenario for this test is that you have an iframe with content from a foreign domain. In that foreign content is an iframe which loads a data: URL. This tests that the data: URL loaded iframe does not have access to the main frame using top.document.


Pass: Cross frame access from a data: URL on a different domain was denied.

--------
Frame: 'aFrame'
--------
Inner iframe on a foreign domain.



--------
Frame: 'aFrame'
--------
Inner-inner iframe. This iframe (which is data: URL and whose parent is on a foreign domain) is the frame attempting to access the main frame. It should not have access to it.
